Add Signin component and route
Add a component and route for signing in via the `/signin` API endpoint.

This commit *does not* implement authenticated Axios connections. While
the response from a successful signin *will* set the csrf token in
localStorage and jwt_access token as a cookie, nothing is actually done
with them yet.

This commit:
* Adds a Signin component and endpoint
* Implements logic to set localStorage on successful signins
  - csrf token
  - signedIn boolean
* Updated Navbar logic for user menus to use `localStorage.signedIn`