Implement addBook method on Books component
Implemented an addBooks method and form on the Books component. This
allows logged in users to add Books to their collections.

The current implementation populates the Authors select field by pulling
the list of all Authors and storing it as a data prop. This might work
well for a small value of authors, but will need a more performant
implementation (likely some form of typeahead) for productionloads. This
would also require a typeahead endpoint on the API server, so we aren't
going to implement it in the Vue app yet.

This commit:
* Adds a form for adding new Books to the Book component
* Places the form in a collapsible div
* Establishes some patterns that will be useful for other POST/PUT/DELETE
  operations