Implement Signup component
The Signup component implements new user signup. Not much to explain.

This commit also moves the Signin component to an explicit `/signin`
path, sets up the Navbar links for signup/signin, and sets the root path
to the Books component. Because the Books component is authenticated,
users are automatically redirected to the signin page if they are not
signed in.