Show Books by Author on Authors component
When you select an Author from the Authors component, a list of books
you own by that Author should be shown. This list should update when you
click a different Author.

This commit adds a list-group to the Authors component, along with a
method selectAuthor() and data props `selectedAuthor` and
`selectedAuthorBooks`.

This functionality requires that the /api/v1/authors/:id/books endpoint
exist on the API server.