Disable webpack host check
After updating webpack-dev-server, web clients were receiving an
'Invalid Host/Origin header' error approximately once per second.

While
the functionality of the application appeared intact, this might cause
unseen errors in the dev test environment and litters the console with
spurious errors, making debugging harder.

This commit disables host checking on webpack-dev-server in the dev
environment. Because it isn't part of the production build chain, it
shouldn't affect security of production servers (though we have yet to
actually run anything in production at this early stage, so we can't be
sure if everything works in that environment yet).