Implement JWT refresh
With the RefreshController implemented on the back-end server we can now
implement refreshing expired access tokens in the front-end
application..

JWT access tokens have a 1-hour expiration time, with the capability to
refresh the token over a 2-week period. In theory this could be
implemented with a separate refresh token, but the current back-end
simply uses the (recently expired) access token to receive a fresh one
with a new 1-hour expiration window.

The refresh action is implemented by tapping into the axios intercept of
the back-end response and looking for 401 Authentication errors. Instead
of immediately redirecting to the signin page, a refresh is attempted.

* If the refresh is successful, the previously unauthenticated action is
  reattempted.
* If the refresh fails, the user is redirected to the login page.

Note: This has not been tested with loops of 401 errors --- that is,
receiving a 401 for a reason *besides* not being logged in or having an
expired token. This needs to be investigated.