Extract flash messages to component
Here we extract all of the display details for flash message to a
separate component from the top-level App component.

We are making a logical distinction behind the *display* details
(including how to respond to the user interacting with flash messages),
and the *message storage* details themselves. The App component will
continue to be the canonical storage location for flash messages, and
thus will handle adding/removing the message objects (and potentially,
storing them in a more permanent location to keep track of message
history).

We also continue the transition in terminology from a singleton `error`
to the pluralized `flashes`. We also make use of a couple of different
flash variants, including the 'info' and 'warning' variants, in addition
to the 'error variant.

The `addError` method remains as an alternative to the `addFlash`
method. In addition to a different default variant, the `addError`
method includes some parsing of the passed-in error object to extract
the actual response from the back-end server, which then gets passed on
to `addFlash` as the `details` object. The `addFlash` method
expects either a basic string or an object representing the item that's
been operated on as it's `details` parameter.