Implement adding Books by ISBN
This feature allows a user to enter the ISBN of a Book, click a Lookup
button, and have the NewBook form filled out automatically. If the
Author of the Book already exists, it should be selected from the
Authors select input. If not, an option to create a new Author should be
selected from the Authors input, with a field for the new Author's name
shown and filled in from the lookup.

The implementation modifies the following:

* Adds new elements to the NewBook form
  - "Create a new Author" in the Authors select box
  - Some convenient help text (disabled) also in the Authors select box
  - "ISBN", "Year", and "New Author Name" inputs
    + The ISBN input has a button attached to perform a lookup. It can
      also be used as just a basic input
    + The New Author Name input only appears if the appropriate option
      is selected in the Author selection
* Wraps the `addBook` form method in a `processForm` method. This method
  first checks if a new Author is to be created
* Adds the `addAuthorThenBok` method. If a new Author is created, this
  method handles creating the new Author, and wraps a call to `addBook`
  in an appropriate `then` block. This is because we need the response
  from the backend containing the new authorId before we can create the
  Book.